backend/requirements.txt
pandas>=1.5
numpy>=1.24
scikit-learn>=1.2
xgboost>=1.7
joblib
Pillow
easyocr

fastapi
uvicorn
python-multipart
tensorflow-cpu
